Added "Court" as an Event type.
I also added as many Event types as we could determine that are in RAM/Sord that aren't in Gedcomx.  This was only meant to start the conversation.  Eric and I sat down and deleted from consideration those that we felt were not appropriate.  We assume you will probably delete/add a few more as you see fit.

Eric also suggested that Randy might want to input on these as well.