Add extension support for Vocab translations, and definitions in VocabConcept
Endpoints are being introduced in Platform API that request, update, or add translations for VocabTerms and VocabConcept definitions. (There is no need for the entire Term or Concept, just the text and associated language.)

Localized definitions were added to Concepts over 2 years ago, but are missing in the VocabConcept model. We need to be able to export this field in order for additional translations to be generated by a partner/vendor.