GedcomxDateRange ends are now inclusive of the specified end date.
For closed ranges:
Treat the unspecified portions of the start date as minimum values .
Treat the unspecified portions of the end date as maximum values.

If the second in the range end date is 59, then add one second to the end date.

The additional second causes the end range to be inclusive and allows the ranges to print cleanly.

This adds the ability to parse range date string of: "+YYYY/+YYYY" where both years are equal.

Equal start and end values that are completely specified down to the second are still invalid as there is no time between the start and end.